requirements.txt
aiogram==2.24
psycopg2-binary==2.9.5
python-dotenv==0.21.0
emoji==2.2.0
requests==2.28.2